Program structure and organization
BATTLEPETS.c is our main function.
BPdesigns are functions that we call for printing the title, dividers, etc designs pretty much.
BPfunctions are the Battle!, ComPetDium, pretty much responsible for the flow.
BPassistfunctions are reusable functions that help the rest of the program.
BPheaders.h contain the defines and function prototypes